fix: make a ClingExecutor to cut down on code reuse
